Fix crashes in the gradient editor
Moving one item exactly to the right of itself is a no-op, but Qt still sends
us a moveRows operation. Then the beginMoveRows function returns false,
meaning this can't/shouldn't be handled, and if we ignore that and call
endMoveRows, Qt crashes.
Detect it and return early. For safety, also test the beginMoveRows return
value, but that alone doesn't fix it, because then the higher level seems to
switch to a backup strategy which also fails - items go missing.