Better Karatsuba multiplication
A special case arises in the Karatsuba algorithm, where (al + ah) is
multiplied by (bl + bh), and the sums are one word wider than their
addends, but the high word is always either zero or one.

If we can detect this, the multiplication can be done more efficiently:
when one of the words involved is known to be either 0 or 1, the
sub-multiply can be replaced by a conditional add.