Emit better code for squaring operations
When computing a*a, it frequently happens that pieces are added in twice.
The algorithm already collects all these and then does a doubling operation
on all of them at once.  However, this precludes using the previous carry
in a multiply-add, it has to be added in afterwards.

In some cases, no doubling is required - for example, if only one piece is
added in the current step, or if everything that would be doubled is zero.
This patch makes us use multiply-add in those cases, removing the need
for an extra addition at the end.