Fix and enable the gen_mult_by_inverse method, for a nice speedup
The problem was that the shift amount was overwritten by emit_normalize_amt.
Easily fixed by copying into a new register.
The fact that we divide by the same number twice in complex division was
the whole reason I tried to implement division by using inversion in the
first place.