Add a "Wind Down" button
Intended to reduce render times for images with a lot of interior points.
It's not always super effective. It might be a good idea to keep track of
maxiter_found in bins and only really increase the value once we exceed
a threshold of maybe 20 pixels in a given bin.