Add several enhancements for Magnet A colouring
Add a formula by David Makin on fractalforums for smooth shading in areas that
don't escape to infinity.

Also add an extra spinbox to add an extra offset to colours in these regions
to make them distinct from the others. Turn the normal offset slider into a
spinbox as well.