Add a few additional colouring options
Add a menu item to make the palette fill the dwell bands (after the transfer
function is applied, as well as width). Makes things line up better with
angle modifiers and allows for certain other effects.  A new option to
make the palette interpolation not circular allows for a repeating b/w
gradient with one sharp edge, for example.